Simplify landing page
