+++ treasury/src/assets/icons/icons.tsx
@@ -5,31 +5,3 @@ type SVGProps = React.SVGAttributes<SVGElement>;
 export interface SVGPropsWithColor extends SVGProps {
   color?: string;
 }
-
-export const AccessIcon: React.FC<SVGPropsWithColor> = ({
-  color,
-  ...props
-}) => (
-  <svg
-    width="19"
-    height="23"
-    viewBox="0 0 19 23"
-    fill="none"
-    xmlns="http://www.w3.org/2000/svg"
-    {...props}
-  >
-    <path
-      fill-rule="evenodd"
-      clip-rule="evenodd"
-      d="M5.4977 4.5C6.04872 4.5 6.49541 4.94669 6.49541 5.4977C6.49541 6.04872 6.04872 6.4954 5.4977 6.4954C4.94669 6.4954 4.5 6.04872 4.5 5.4977C4.5 4.94669 4.94669 4.5 5.4977 4.5ZM6.49544 9.65477C6.49544 9.10375 6.04875 8.65707 5.49774 8.65707C4.94672 8.65707 4.50004 9.10375 4.50004 9.65477C4.50004 10.2058 4.94672 10.6525 5.49774 10.6525C6.04875 10.6525 6.49544 10.2058 6.49544 9.65477ZM14.8096 9.65477C14.8096 9.10375 14.3629 8.65707 13.8119 8.65707C13.2609 8.65707 12.8142 9.10375 12.8142 9.65477C12.8142 10.2058 13.2609 10.6525 13.8119 10.6525C14.3629 10.6525 14.8096 10.2058 14.8096 9.65477ZM10.6525 17.969C10.6525 17.418 10.2058 16.9713 9.65481 16.9713C9.10379 16.9713 8.65711 17.418 8.65711 17.969C8.65711 18.52 9.10379 18.9667 9.65481 18.9667C10.2058 18.9667 10.6525 18.52 10.6525 17.969ZM9.65481 8.65707C10.2058 8.65707 10.6525 9.10375 10.6525 9.65477C10.6525 10.2058 10.2058 10.6525 9.65481 10.6525C9.10379 10.6525 8.65711 10.2058 8.65711 9.65477C8.65711 9.10375 9.10379 8.65707 9.65481 8.65707ZM10.6525 5.4977C10.6525 4.94669 10.2058 4.5 9.65481 4.5C9.10379 4.5 8.6571 4.94669 8.6571 5.4977C8.6571 6.04872 9.10379 6.4954 9.65481 6.4954C10.2058 6.4954 10.6525 6.04872 10.6525 5.4977ZM13.8119 4.5C14.3629 4.5 14.8096 4.94669 14.8096 5.4977C14.8096 6.04872 14.3629 6.4954 13.8119 6.4954C13.2609 6.4954 12.8142 6.04872 12.8142 5.4977C12.8142 4.94669 13.2609 4.5 13.8119 4.5ZM10.6525 13.8119C10.6525 13.2609 10.2058 12.8142 9.65481 12.8142C9.10379 12.8142 8.65711 13.2609 8.65711 13.8119C8.65711 14.3629 9.10379 14.8096 9.65481 14.8096C10.2058 14.8096 10.6525 14.3629 10.6525 13.8119ZM5.49774 12.8142C6.04875 12.8142 6.49544 13.2609 6.49544 13.8119C6.49544 14.3629 6.04875 14.8096 5.49774 14.8096C4.94672 14.8096 4.50004 14.3629 4.50004 13.8119C4.50004 13.2609 4.94672 12.8142 5.49774 12.8142ZM14.8096 13.8119C14.8096 13.2609 14.3629 12.8142 13.8119 12.8142C13.2609 12.8142 12.8142 13.2609 12.8142 13.8119C12.8142 14.3629 13.2609 14.8096 13.8119 14.8096C14.3629 14.8096 14.8096 14.3629 14.8096 13.8119Z"
-      fill={color || "white"}
-    />
-    <path
-      d="M16.037 1H2.87963C1.84154 1 1 2.04467 1 3.33333V19.6667C1 20.9553 1.84154 22 2.87963 22H16.037C17.0751 22 17.9167 20.9553 17.9167 19.6667V3.33333C17.9167 2.04467 17.0751 1 16.037 1Z"
-      stroke={color || "white"}
-      stroke-width="2"
-      stroke-linecap="round"
-      stroke-linejoin="round"
-    />
-  </svg>
-);
